Code formatting/refactoring
Add whitespace
Add CML Discord link
Remove unused variables
Change function expressions to arrow functions
Increase CML logo size
