+++ src/components/layout/layout.tsx
@@ -2,11 +2,13 @@ import Head from "next/head";
 import { CustomHead } from "./customHead";
 import { BackgroundImage, Grid, createStyles } from "@mantine/core";
 import { Header } from "./header";
-import { Navbar } from "./navbar";
+import { Navbar } from "./navbar/navbar";
 import { Footer } from "./footer";
 
 
-const useStyles = createStyles((_) => ({
+
+
+const useStyles = createStyles(() => ({
     backgroundImage: {
         minWidth: "750px",
         minHeight: "100vh",
@@ -18,6 +20,8 @@ const useStyles = createStyles((_) => ({
         padding: "5px 45px",
     },
 }));
+
+
 
 
 export type LayoutProps = {
@@ -32,7 +36,9 @@ export type LayoutProps = {
 };
 
 
-export function Layout({
+
+
+export const Layout = ({
     children,
     pageTitle,
     pageDescription,
@@ -41,53 +47,53 @@ export function Layout({
     robotsText = "index,follow,nositelinkssearchbox",
     socialMediaImageUrl = "https://celestemods.com/img/cml_icon.png",
     socialMediaImageAlt = "Celeste Mods List Logo",
-}: LayoutProps) {
-  const { classes } = useStyles();
+}: LayoutProps) => {
+    const { classes } = useStyles();
 
 
-  return (
-    <>
-        <Head>
-            <CustomHead
-                pageTitle={pageTitle}
-                pageDescription={pageDescription}
-                pathname={pathname}
-                siteName={siteName}
-                robotsText={robotsText}
-                socialMediaImageUrl={socialMediaImageUrl}
-                socialMediaImageAlt={socialMediaImageAlt}
-            />
-        </Head>
-        <BackgroundImage
-            src="/images/cml_background1.png"
-            className={classes.backgroundImage}
-        >
-            <Grid gutter="0">
-                <Grid.Col span={2}></Grid.Col>
-                <Grid.Col span={8}>
-                    <Header />
-                </Grid.Col>
-                <Grid.Col span={2}></Grid.Col>
-                <Grid.Col span={2}>
-                    <Navbar
-                        pathname={pathname}
-                        pages={[
-                            { label: "Home", pathname: "/" },
-                            { label: "Mods", pathname: "/mods" },
-                        ]}
-                    />
-                </Grid.Col>
-                <Grid.Col span={8}>
-                    <main className={classes.children}>{children}</main>
-                </Grid.Col>
-                <Grid.Col span={2}></Grid.Col>
-                <Grid.Col span={2}></Grid.Col>
-                <Grid.Col span={8}>
-                    <Footer />
-                </Grid.Col>
-                <Grid.Col span={2}></Grid.Col>
-            </Grid>
-        </BackgroundImage>
-    </>
-  );
-}
+    return (
+        <>
+            <Head>
+                <CustomHead
+                    pageTitle={pageTitle}
+                    pageDescription={pageDescription}
+                    pathname={pathname}
+                    siteName={siteName}
+                    robotsText={robotsText}
+                    socialMediaImageUrl={socialMediaImageUrl}
+                    socialMediaImageAlt={socialMediaImageAlt}
+                />
+            </Head>
+            <BackgroundImage
+                src="/images/cml_background1.png"
+                className={classes.backgroundImage}
+            >
+                <Grid gutter="0">
+                    <Grid.Col span={2}></Grid.Col>
+                    <Grid.Col span={8}>
+                        <Header />
+                    </Grid.Col>
+                    <Grid.Col span={2}></Grid.Col>
+                    <Grid.Col span={2}>
+                        <Navbar
+                            pathname={pathname}
+                            pages={[
+                                { label: "Home", pathname: "/" },
+                                { label: "Mods", pathname: "/mods" },
+                            ]}
+                        />
+                    </Grid.Col>
+                    <Grid.Col span={8}>
+                        <main className={classes.children}>{children}</main>
+                    </Grid.Col>
+                    <Grid.Col span={2}></Grid.Col>
+                    <Grid.Col span={2}></Grid.Col>
+                    <Grid.Col span={8}>
+                        <Footer />
+                    </Grid.Col>
+                    <Grid.Col span={2}></Grid.Col>
+                </Grid>
+            </BackgroundImage>
+        </>
+    );
+};
